dissertation_data

This dataset, curated specifically for training models in my dissertation, encompasses variables from the YRBS 2017, 2019, and 2021 datasets. Additionally, several variables have been specifically constructed by combining common variables; for example, 'RiskDrivingBehaviors' is a composite variable derived from 'SeatBeltUse', 'TextingDriving', 'DrivingDrinking', and 'DrinkingDriver'. The selection of variables is strategically focused on analyzing suicidality Rigorous curation ensured that each included variable has less than 80% missing data. The selection prioritizes variables that are pertinent to the analysis of suicidality. For details on the exclusion criteria and a list of removed variables, refer to 'inst/scripts/creatingCompleteDataset.qmd'. NOTE: Missing data for SuicideAttempts

#' This dataset exclusively comprises questions sourced from Appendix A of the YRBS 2017, 2019, and 2021, specifically focusing on the QN dichotomous variables. For comprehensive information, please refer to the documentation provided in the 2017, 2019, and 2021 National YRBS Data Users Guide. In this dataset, each variable follows a binary pattern: a value of 1 indicates the presence of the specified behavior, while a value of 0 signifies its absence. For instance, with regard to Suicide Attempts, a value of 1 signifies that adolescents responded affirmatively to having made suicide attempts, whereas 0 indicates a negative response.

Variables in the dataset include:

year

Survey year (e.g., 2017, 2019, 2021).

weight

Weight variable for survey adjustment.

Sex

Participant's sex (Male/Female).

Age

Participant's age.

Grade

School grade (9th–12th).

SexOrientation

Sexual orientation of the participant.

WeaponCarryingSchool

Carrying a weapon on school property.

GunCarrying

Carrying a gun.

UnsafeAtSchool

Feeling unsafe at school.

ThreatenedInSchool

Being threatened at school.

ForcedSexualIntercourse

Having been forced into sexual intercourse.

PhysicalDaitingViolence

Experiencing physical dating violence.

Bullying

Being bullied.

CyberBullying

Being cyberbullied.

Hopelessness

Feeling hopeless for 2+ weeks.

SuicideIdeation

Considering suicide.

SuicidePlan

Making a suicide plan.

SuicideAttempts

Attempting suicide.

CurrentlySmokingCigarette

Currently smoking cigarettes.

EverTriedVaping

Ever tried vaping.

CurrentlyVaping

Currently vaping.

CurrentlySmokelessTobacco

Currently using smokeless tobacco.

CurrentlySmokingCigar

Currently smoking cigars.

QuitTobbaco

Attempted to quit tobacco use.

FirstAlcoholBefore13

Had first drink of alcohol before age 13.

CurrentlyAlcohol

Currently drinking alcohol.

CurrentlyBingeDrinking

Currently binge drinking.

EverUsedMarihuana

Ever used marijuana.

FirstMarihuanaBefore13

Used marijuana before age 13.

CurrentlyUseMarihuana

Currently using marijuana.

EverUsedSyntheticMarihuana

Ever used synthetic marijuana.

PainMedicine

Misusing prescription pain medicine.

EverUsedInjectedIllegalDrug

Ever injected illegal drugs.

OfferedDrugsSchool

Offered drugs on school property.

EverHadSex

Ever had sexual intercourse.

SexBefore13

Had sexual intercourse before age 13.

Sex4OrMorePartners

Had 4 or more sexual partners.

SexuallyActive

Currently sexually active.

VeryOverweight

Perceived as very overweight.

NoSoda

Does not drink soda.

NoMilk

Does not drink milk.

NoBreakfast

Does not eat breakfast.

ThreeOrMoreHoursTV

Spends 3+ hours watching TV.

ThreeOrMoreHoursVideoGames

Spends 3+ hours on video games.

AttendedPEClass

Attends physical education classes.

ConcussionSports

Experienced a concussion during sports.

EightorMoreHoursSleep

Gets 8+ hours of sleep.

NoSportsDrinks

Does not consume sports drinks.

NoDrinksWater

Does not drink water.

DifficultyConcentrating

Has difficulty concentrating.

RaceEthnicity

Race/ethnicity of the participant.

HardDrugsUse

Composite variable for hard drug use.

SportAndActivity

Composite variable for sports and physical activities.

RiskDrivingBehaviors

Composite variable for risky driving behaviors.

SexualAbuse

Composite variable for sexual abuse.

TriedCigarrete

Composite variable for trying cigarettes.

NoFruitOrVeggies

Composite variable for not eating fruits or vegetables.

PhysicalFights

Involvement in physical fights.

data(dissertation_data)

Format

A tibble with 36779 rows and 58 variables:

Details

Dataset with Suicide Related Predictors Using the 2017, 2019 and 2021 YRBSS data