raw2019

This is the imported raw version of the 2019 YRBS Data. This data set monitors priority health risk behaviors that contribute markedly to the leading causes of death, disability, and social problems among youth and adults in the United States. These behaviors, often include:

#' Please refer to the documentation provided in the 2019 National YRBS Data Users Guide.

These broad categories include:

Behaviors that contribute to unintentional injuries and violence

Sexual behaviors that contribute to unintended pregnancy and sexually transmitted diseases, including HIV infection

Alcohol and other drug use

Tobacco use

Unhealthy dietary behaviors

Inadequate physical activity

data(raw2019)

Format

A tibble with 13,677 and 235 variables

Details

Youth Risk Behavior Surveillance System 2019