dissertation_data

This dataset, curated specifically for training models in my dissertation, encompasses variables from the YRBS 2017, 2019, and 2021 datasets. Additionally, several variables have been specifically constructed by combining common variables; for example, 'RiskDrivingBehaviors' is a composite variable derived from 'SeatBeltUse', 'TextingDriving', 'DrivingDrinking', and 'DrinkingDriver'. The selection of variables is strategically focused on analyzing suicidality Rigorous curation ensured that each included variable has less than 80% missing data. The selection prioritizes variables that are pertinent to the analysis of suicidality. For details on the exclusion criteria and a list of removed variables, refer to 'inst/scripts/creatingCompleteDataset.qmd'.

year

weight

Sex

Age

Grade

SexOrientation

WeaponCarryingSchool

GunCarrying

UnsafeAtSchool

ThreatenedInSchool

ForcedSexualIntercourse

PhysicalDaitingViolence

Bullying

CyberBullying

Hopelessness

SuicideIdeation

SuicidePlan

SuicideAttempts

CurrentlySmokingCigarette

EverTriedVaping

CurrentlyVaping

CurrentlySmokelessTobacco

CurrentlySmokingCigar

QuitTobbaco

FirstAlcoholBefore13

CurrentlyAlcohol

CurrentlyBingeDrinking

EverUsedMarihuana

FirstMarihuanaBefore13

CurrentlyUseMarihuana

EverUsedSyntheticMarihuana

PainMedicine

EverUsedInjectedIllegalDrug

OfferedDrugsSchool

EverHadSex

SexBefore13

Sex4OrMorePartners

SexuallyActive

VeryOverweight

NoSoda

NoMilk

NoBreakfast

ThreeOrMoreHoursTV

ThreeOrMoreHoursVideoGames

AttendedPEClass

ConcussionSports

EightorMoreHoursSleep

NoSportsDrinks

NoDrinksWater

DifficultyConcentrating

RaceEthnicity

HardDrugsUse composite variable

SportAndActivity composite variable

RiskDrivingBehaviorsType label here

SexualAbuse composite variable

TriedCigarrete composite variable

NoFruitOrVeggies composite variable

PhysicalFights PhysicalFights

data(dissertation_data)

Format

A tibble with 36779 rows and 58 variables:

This dataset exclusively comprises questions sourced from Appendix A of the YRBS 2017, 2019 and 2021, specifically focusing on the QN dichotomous variables. For comprehensive information, please refer to the documentation provided in the 2017, 2019 and 2021 National YRBS Data Users Guide. In this dataset, each variable follows a binary pattern: a value of 1 indicates the presence of the specified behavior, while a value of 0 signifies its absence. For instance, with regard to Suicide Attempts, a value of 1 signifies that adolescents responded affirmatively to having made suicide attempts, whereas 0 indicates a negative response.

Details

Dataset with Suicide Related Predictors Using the 2017, 2019 and 2021 YRBSS data